requirements.txt
google-auth==0.7.0
google-auth-httplib2==0.0.2
google-cloud-bigquery==0.22.1
google-cloud-core==0.22.1
googleapis-common-protos==1.5.2
httplib2==0.10.3
numexpr==2.6.2
numpy==1.12.0
pandas==0.19.2
protobuf==3.2.0
pyasn1==0.2.2
pyasn1-modules==0.0.8
python-dateutil==2.6.0
pytz==2016.10
rsa==3.4.2
six==1.10.0
tables==3.3.0
